Issue #13 - Refactored the mount operations to a vfsmount.h file
