+++ src/Makefile
@@ -29,6 +29,7 @@ OBJS = \
 	uart.o\
 	vectors.o\
 	vfs.o\
+	vfsmount.o\
 	vm.o\
 
 # Cross-compiling (e.g., on Mac OS X)
