+++ src/fs.h
@@ -1,6 +1,8 @@
-// On-disk file system format. 
+// On-disk file system format.
 // Both the kernel and user programs use this header file.
 
+#ifndef XV6_FS_H_
+#define XV6_FS_H_
 
 #define ROOTINO 1  // root i-number
 #define BSIZE 512  // block size
@@ -54,3 +56,4 @@ struct dirent {
   char name[DIRSIZ];
 };
 
+#endif /* XV6_FS_h */
