+++ src/main.c
@@ -1,5 +1,6 @@
 #include "types.h"
 #include "defs.h"
+#include "vfsmount.h"
 #include "param.h"
 #include "memlayout.h"
 #include "mmu.h"
